Use table for videos list, refs #753
