+++ vueapp/views/VideosTrashed.vue
@@ -1,17 +1,17 @@
 <template>
     <div>
-        <VideosList></VideosList>
+        <VideosTable></VideosTable>
     </div>
 </template>
 
 <script>
 import { mapGetters } from "vuex";
-import VideosList from "@/components/Videos/VideosList";
+import VideosTable from "@/components/Videos/VideosTable";
 
 export default {
     name: "VideosArchive",
     components: {
-        VideosList
+      VideosTable
     }
 };
 </script>
